+++ src/main/java/com/gregori/order/mapper/OrderMapper.java
@@ -1,0 +1,13 @@
+package com.gregori.order.mapper;
+
+import java.util.Optional;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import com.gregori.order.domain.Order;
+
+@Mapper
+public interface OrderMapper {
+	Long insert(Order order);
+	Optional<Order> findById(Long orderId);
+}
