+++ src/main/java/com/gregori/order/service/OrderService.java
@@ -1,0 +1,9 @@
+package com.gregori.order.service;
+
+import com.gregori.order.dto.OrderRequestDto;
+import com.gregori.order.dto.OrderResponseDto;
+
+public interface OrderService {
+	Long createOrder(OrderRequestDto orderRequestDto);
+	OrderResponseDto findOrderById(Long orderId);
+}
