--- src/main/java/com/gregori/service/order/OrderService.java
@@ -1,9 +1,0 @@
-package com.gregori.service.order;
-
-import com.gregori.dto.order.OrderRequestDto;
-import com.gregori.dto.order.OrderResponseDto;
-
-public interface OrderService {
-	Long createOrder(OrderRequestDto orderRequestDto);
-	OrderResponseDto findOrderById(Long orderId);
-}
